small enough to inline
